Refactor Toggle-cell to be (more) reusable (IOS-14)
